move the versioning functions to pgsql.common
